refactor: Update fetchUtils import paths to use the correct directory structure
