+++ components/dashboard/attendanceTable.tsx
@@ -3,7 +3,7 @@ import AttendanceTableBody from '@/components/dashboard/attendanceTableBody'
 import { AttendanceProvider } from '@/app/context/attendanceContext'
 import { getSession, Session } from '@auth0/nextjs-auth0'
 import { UserInfoProps } from '@/types/interfaces'
-import { fetchUserInfoByEmail } from '@/app/utils/fetchUtils'
+import { fetchUserInfoByEmail } from '@/utils/fetchUtils'
 import AttendanceTableRefresh from '@/components/dashboard/attendanceTableRefresh'
 
 export default async function AttendanceTable() {
